add: list operations

+++ src/models/List.js
@@ -1,4 +1,4 @@
-import {gql} from 'apollo-server'
+import {ApolloError, AuthenticationError, gql} from 'apollo-server'
 import {isAuthenticated} from '../auth';
 import List from '../database_models/List'
 import User from '../database_models/User';
@@ -10,7 +10,7 @@ export const typeDefs = gql`
   extend type Mutation {
     createList(title: String!): CreateListMutationResponse!
     deleteList(listId: ID!): DeleteListMutationResponse!
-    addListItem(listId: ID!, title: String!, url: String!): AddListItemMutationResponse!
+    addListItem(listId: ID!, title: String!, url: String!, externalId: String!, category: String!): AddListItemMutationResponse!
     deleteListItem(listId: ID!, listItemId: ID!): DeleteListItemMutationResponse!
   }
   type List {
@@ -23,6 +23,8 @@ export const typeDefs = gql`
     id: ID!
     title: String!
     url: String!
+    category: String!
+    externalId: String!
   }
   type CreateListMutationResponse implements MutationResponse{
     code: String!
@@ -59,7 +61,7 @@ export const resolvers = {
   },
   Mutation: {
     createList: isAuthenticated(async (_, {title}, {user}) => {
-      const list = await List.findOne({title});
+      const list = await List.findOne({title}).lean().exec();
       if (list) {
         return {
           code: "404",
@@ -78,33 +80,50 @@ export const resolvers = {
         list: newList
       };
     }),
-    deleteList: async (_, {listId}) => {
+    deleteList: isAuthenticated(async (_, {listId}, {user}) => {
+      const list = await List.findById(listId);
+      if (!list) throw new ApolloError("List Not Found")
+      if (list.userId.toString() !== user._id.toString()) {
+        throw new AuthenticationError(`This list doesn't belong to you`);
+      }
       await List.findByIdAndDelete(listId);
       return {
         code: "202",
         success: true,
         message: "list deleted successfully"
       }
-    },
-    addListItem: async (_, {listId, title, url}) => {
-      const list = await List.findByIdAndUpdate(listId, {
+    }),
+    addListItem: isAuthenticated(async (_, {listId, title, url, category, externalId}, {user}) => {
+      const list = await List.findById(listId);
+      if (!list) throw new ApolloError("List Not Found");
+      if (list.userId.toString() !== user._id.toString()) {
+        throw new AuthenticationError(`This list doesn't belong to you`);
+      }
+      const updatedList = await List.findByIdAndUpdate(listId, {
         $push: {
           listItems: {
-            title, url
+            title, url, category, externalId
           }
         }
       }, {new: true});
-      const newListItem = list.listItems[list.listItems.length - 1];
+      const newListItem = updatedList.listItems[updatedList.listItems.length - 1];
       return {
         code: "201",
         success: true,
         message: "new listItem created",
-        list: list,
+        list: updatedList,
         listItem: newListItem
       }
-    },
-    deleteListItem: async (_, {listId, listItemId}) => {
-      const list = await List.findByIdAndUpdate(listId, {
+    }),
+    deleteListItem: isAuthenticated(async (_, {listId, listItemId}, {user}) => {
+      const list = await List.findById(listId);
+      if (!list) {
+        throw new ApolloError("List Not Found")
+      }
+      if (list.userId.toString() !== user._id.toString()) {
+        throw new AuthenticationError(`This list doesn't belong to you`);
+      }
+      const updatedList = await List.findByIdAndUpdate(listId, {
         $pull: {
           listItems: {
             _id: listItemId
@@ -122,9 +141,9 @@ export const resolvers = {
         code: "202",
         success: true,
         message: "listitem delted successfully",
-        list
+        list: updatedList
       };
-    }
+    })
   },
   List: {
     user: async (list) => {
